+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix: Do not zoom to error when geometry is null geometry
+
 ## [1.1.4] - 2023-03-08
 
 - Fix: Show correct error count when errors are filtered